% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2canvas.R
\name{html2canvas}
\alias{html2canvas}
\title{<Add Title>}
\usage{
html2canvas(
  id,
  label,
  selector,
  background = "black",
  color = "white",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
